Missing File
Core parts of Maralook depend on this file. It currently is missing in my last backup. Will look for it and if needed re-write the file. It mainly consisted of my own version of the item information marapets shows when clicking on "check price" or whatever. A very critical file for Maralook..